replacing explicit trig float functions with normal ones

+++ DirectionalCompressor/Source/PluginEditor.cpp
@@ -348,8 +348,8 @@ void AmbisonicCompressorAudioProcessorEditor::IEMSphereElementChanged (IEMSphere
     float hypxy = sqrt(pos.x*pos.x+pos.y*pos.y);
     
     
-    float yaw = atan2f(pos.y,pos.x);
-    float pitch = atan2f(hypxy,pos.z)-M_PI/2;
+    float yaw = atan2(pos.y,pos.x);
+    float pitch = atan2(hypxy,pos.z)-M_PI/2;
     
     valueTreeState.getParameter("yaw")->setValue(valueTreeState.getParameterRange("yaw").convertTo0to1(yaw/M_PI*180.0f));
     valueTreeState.getParameter("pitch")->setValue(valueTreeState.getParameterRange("pitch").convertTo0to1(pitch/M_PI*180.0f));
